requirements.txt
flask==3.0.3
flask-limiter
pillow
numpy
torch
transformers